refactor(frontend/context): move context folder out of app
