+++ frontend/context/appletContext.tsx
